refactor: UserRequested 객체의 이름을 UserRequestedParticipants 로 수정한다
