+++ src/main/java/laddergame/domain/game/UserRequestedParticipants.java
@@ -1,20 +1,20 @@
 package laddergame.domain.game;
 
-public class UserRequest {
+public class UserRequestedParticipants {
 
     public static final String ALL_PARTICIPANTS_COMMAND = "all";
-    private static final UserRequest ALL_PARTICIPANTS = new UserRequest(ALL_PARTICIPANTS_COMMAND);
+    private static final UserRequestedParticipants ALL_PARTICIPANTS = new UserRequestedParticipants(ALL_PARTICIPANTS_COMMAND);
     private final String requestContent;
 
-    private UserRequest(final String requestContent) {
+    private UserRequestedParticipants(final String requestContent) {
         this.requestContent = requestContent;
     }
 
-    public static UserRequest from(final String requestContent) {
+    public static UserRequestedParticipants from(final String requestContent) {
         if (requestContent.equals(ALL_PARTICIPANTS_COMMAND)) {
             return ALL_PARTICIPANTS;
         }
-        return new UserRequest(requestContent);
+        return new UserRequestedParticipants(requestContent);
     }
 
     public boolean isAllParticipants() {
