+++ src/test/java/laddergame/domain/game/LadderGameTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import java.util.List;
 import java.util.Map;
 
-import static laddergame.domain.game.UserRequest.ALL_PARTICIPANTS_COMMAND;
+import static laddergame.domain.game.UserRequestedParticipants.ALL_PARTICIPANTS_COMMAND;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class LadderGameTest {
@@ -40,7 +40,7 @@ class LadderGameTest {
     @CsvSource(value = {"one:1st", "two:2nd", "three:3rd", "four:4th"}, delimiter = ':')
     @DisplayName("참여자 이름을 입력하면, 한 명의 참여자에 대한 실행결과를 얻는다.")
     void gets_result_of_participant_if_request_content_is_included_in_participants(String requestContent, String expectedResult) {
-        UserRequest request = UserRequest.from(requestContent);
+        UserRequestedParticipants request = UserRequestedParticipants.from(requestContent);
 
         Map<Participant, Result> resultByParticipants = ladderGame.getResultByParticipants(request);
         String actualResult = getActualResult(requestContent, resultByParticipants);
@@ -53,7 +53,7 @@ class LadderGameTest {
     void gets_all_results_if_request_content_is_same_with_all_request_key() {
         List<Participant> allParticipants = participants.getParticipants();
         List<String> resultNames = List.of("1st", "2nd", "3rd", "4th");
-        UserRequest request = UserRequest.from(ALL_PARTICIPANTS_COMMAND);
+        UserRequestedParticipants request = UserRequestedParticipants.from(ALL_PARTICIPANTS_COMMAND);
 
         Map<Participant, Result> resultByParticipants = ladderGame.getResultByParticipants(request);
